fixing docker-compose

+++ srcs/requirements/wordpress/tools/install.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
-wget https://wordpress.org/latest.tar.gz
-tar -xf latest.tar.gz
-mv latest.tar.gz wordpress/* /var/www/html
-mv wp-config.php /var/www/html
+if [ ! -f /var/www/html/wp-config.php ]
+then
+	wget https://wordpress.org/latest.tar.gz
+	tar -xf latest.tar.gz
+	mv latest.tar.gz wordpress/* /var/www/html
+	mv wp-config.php /var/www/html
 
-sed -i "s/database_name_here/$WORDPRESS_DATABASE_NAME/" /var/www/html/wp-config.php
-sed -i "s/username_here/$WORDPRESS_DATABASE_USER/" /var/www/html/wp-config.php
-sed -i "s/password_here/$WORDPRESS_DATABASE_PASSWORD/" /var/www/html/wp-config.php
+	sed -i "s/database_name_here/$WORDPRESS_DATABASE_NAME/" /var/www/html/wp-config.php
+	sed -i "s/username_here/$WORDPRESS_DATABASE_USER/" /var/www/html/wp-config.php
+	sed -i "s/password_here/$WORDPRESS_DATABASE_PASSWORD/" /var/www/html/wp-config.php
+	sed -i "s/localhost/ma/" /var/www/html/wp-config.php
+fi
 
 #wp core install \
 #  --url="http://ael-mhar.42.fr" \
@@ -16,4 +20,4 @@ sed -i "s/password_here/$WORDPRESS_DATABASE_PASSWORD/" /var/www/html/wp-config.p
 #  --admin_password="$WORDPRESS_PASSWORD" \
 #  --admin_email="$WORDPRESS_EMAIL"
 
-php-fpm8.2
+php-fpm8.2 -F
